TuneableConsistencyScatter
